+++ src/ladb_opencutlist/ruby/worker/common/common_drawing_projection_worker.rb
@@ -226,7 +226,7 @@ module Ladb::OpenCutList
 
       end
 
-      # -- Shapes
+      # -- Shell
 
       if @compute_shell